feat: Adicionando validaçcao no metodo de delete de lista de favoritos. Apenas o próprio usuário pode deletar a própria lista. Ajustando algumas entradas de informação no data.sql. Adicionando um usuario padrão com idade de 15.

+++ serratecFlix/src/main/java/org/serratec/serratecFlix/service/ListaFavoritosService.java
@@ -131,9 +131,18 @@ public class ListaFavoritosService {
         return new ListaFavoritosResponseDTO(lista);
     }
 
-    public void deletar(Long id) {
-        listaFavoritosRepository.findById(id)
+    public void deletar(Long id, String username) {
+        ListaFavoritos lista = listaFavoritosRepository.findById(id)
                 .orElseThrow(() -> new ValorNaoEncontradoException("Não encontramos uma Lista de Favoritos com esse identificador."));
+
+        Usuario usuario = usuarioRepository.findByUsername(username);
+        if (usuario == null) {
+            throw new ValorNaoEncontradoException("Usuario não encontrado");
+        }
+
+        if (!lista.getUsuario().getId().equals(usuario.getId())) {
+            throw new ValorNaoEncontradoException("Você não pode excluir a lista de favoritos de outro usuário.");
+        }
 
         listaFavoritosRepository.deleteById(id);
     }
